feat: Empty GLFW window without OpenGL context
To prepare for Vulkan initialization we need a GLFW window that has no
OpenGL context created for it. Because GLFW was created with OpenGL in
mind we need to provide a `NoApi` window hint.

```
glfw.window_hint(glfw::WindowHint::ClientApi(glfw::ClientApiHint::NoApi));
```

For convenience we also handle the press of the `Esc` key to close the
window.