fix: Swapchain semaphore reuse
The validation layers were crying about queue signal semaphores not
being available or something like this. The validation in question is
`VUID-vkQueueSubmit-pSignalSemaphores-00067`.

The solution is to create as many wait semaphores as there are swapchain
images and index into this array using the image index returned by
`vkAcquireNextImageKHR` instead of the index of the current in flight
frame.

See:
 -
https://registry.khronos.org/vulkan/specs/latest/man/html/vkQueueSubmit.html#VUID-vkQueueSubmit-pSignalSemaphores-00067
- https://docs.vulkan.org/guide/latest/swapchain_semaphore_reuse.html