feat: Descriptor pool and sets
This commit actually binds the uniform buffer and the shader is able to
use it.

Honestly after this commit I'm in the team why the fuck is Vulkan so
convoluted.

One thing is for sure. I need to reimplement this from scrats for it to
cement in my brain, at least conceptually.