fix: Flip Y axis and set front face to counter clock wise
Because `glm` is meant to work with OpenGL we have to flip the Y
component of the projection matrix.

When doing so the trangles will get culled because they will have their
back side facing the camera. To make them face the camera we had to
change the definition of a front face from clock wise to counter clock
wise.