feat: Load OBJ / Waveform model
Use `tobj` create to load the model.

This commit has a bug, the 3D mesh is loaded "exploded" it looks like
wrong position or wrong indices, but I'll focus in another commit to do
a fix.

For now I'm happy that I load the texture correctly. I don't know if
texture coordinates are good because the mesh is not correctly loaded.