feat: Adding new attachments
To be honest I don't understand right now how all this MSAA business
works especially when it comes to attachments. :-s

Confused as fuck, but at least I learned a hard lesson. Make sure you
create resources with the correct sizes. I was creating the MSAA color
image with the initial with and height of the image, not the ones from
the swapchain extent. It was painful to track down as I did not the
setup to attach a debugger and I'm not used to the Vulkan validation
errors.