feat(log): Nove logging related functions in the `log` module
Moved the `print_error` function to the `log` module, but also renamed
it to `error_details`. In this way it looks better when we call it like
this: `log::error_details(some_anyhow_error)`.

By introducing the `log` module we no longer have access to the `log`
create macros, therever in our own `log` module we re-export everything
from the `log` crate.

Also the reasons the error occured are now printed in reverse order.