refactor(engine): Move `log` module to `engine` crate
It's time to split off engine elements in the `engine` crate and
possibly other crates.

For starters the `log` module is a good thing to start with.

At the moment I'm not quite sure how I'll do it, but I expect the
`engine` crate to contain all the platform and rendering hardware
agnostic elements, while different implementations will live in a crate
with a name like `engine_vulkan` or `engine_platform_linux` etc.

I will see where the flow takes me as this is a new thing for me. I'm
new at abstracting things in Rust and especially abstracting rendering
APIs.