fix(vk-sbox): Load resources in a save manner
First we try to load the resources from `CARGO_MANIFEST_DIR`, if we
don't find the assets there we try to load them from the current working
directory. This helps with being able to run the executable using
`cargo run`, but also it will be fine when distributing the binary and
the assets.